# Disable nop padding in R2R images and remove jumpstamps from the runtime