… clrdefinitions.cmake
… src/inc/readytorun.h
… src/vm/.vscode/c_cpp_properties.json
… src/vm/methodtable.cpp
… src/vm/prestub.cpp
… src/vm/rejit.cpp
… src/zap/zapinfo.cpp