This is still a work in progress, there are some failing tests and I still need to run perf tests on it, but I want to get eyes on it as early as possible.


This change does three things:

- Remove nop padding from R2R images
- Remove jumpstamps from the runtime and switch rejit to use exclusively precodes
- Merge my changes with Kount's recent changes in code versioning

Remaining work before it's ready to merge:

- [x] Get all tests passing
- [ ] Hopefully refactor ReJitManager::UpdateActiveILVersions so it no longer requires a runtime suspension (moved to a separte issue #26790)
- [x] Run perf tests on it to prove there are no startup regressions