@jkotas 2019-09-17
This change needs `READYTORUN_MAJOR_VERSION` bump since it changes the contract of precompiled code. The existing runtimes are not able to run the new code anymore.