@davmason 2019-09-17
@jkotas I don't need to bump `MINIMUM_READYTORUN_MAJOR_VERSION` though, right? Since the runtime can still use the images with nop padding in them, it's just old runtimes can't use the new ones.