@jkotas 2019-09-17
> I don't need to bump MINIMUM_READYTORUN_MAJOR_VERSION though, right?

Right.