@davmason 2019-09-17
@jkotas @noahfalk I was running in to failures where we would have an R2R image with methods that had precodes baked in but not native code slots. When tests disabled tiered compilation and rejit (not a lot, but there are a few) initial jitting would fail for these methods because we expect if there is a precode we also have a native code slot. My latest commit https://github.com/dotnet/coreclr/pull/26740/commits/e20fac27f8f4bf066b12e1416a77e172b96fb14e changes it so if we generate a precode in a saved MethodDesc we also generate a saved native code slot.

I haven't been able to convince myself that this is always the right thing to do, but it does make the tests pass. I'd appreciate any input on this.