@kouvel 2019-09-18
> failures where we would have an R2R image with methods that had precodes baked in but not native code slots

I wasn't aware that some MethodDescs are baked into R2R images, I thought that was only done for NGen and was avoided in R2R to have better versionability, though it would be reasonable for CoreLib methods. What type of methods are these?

> if we generate a precode in a saved MethodDesc we also generate a saved native code slot

I don't see an issue with adding a native code slot other than some size increase. I guess it would depend on what information is available at the time, if it's possible to check the preconditions that would satisfy versionability excluding run-time components then it could be limited to only those cases. My understanding is that there are many types of methods that get precodes and are not versionable (stubs for instance), so not sure about the extent of the issue.