@jkotas 2019-09-18
> I wasn't aware that some MethodDescs are baked into R2R images, I thought that was only done for NGen and was avoided in R2R to have better versionability,

Correct. MethodDescs are never baked into R2R images, not even for CoreLib.