@davmason 2019-09-19
I just removed the draft status, I was originally intending to do some more work and hopefully remove the runtime suspension during a ReJIT, but I think this change is big enough and complicated enough on its own. Filed #26790 to track the additional work.