@davmason 2019-09-27
I ran startup perf tests and see no regressions (details below). At this point I have responded to all feedback and this is in a state where I feel confident checking it in. If there are any outstanding issues, please let me know so I can address them.

cc @brianrob @billwert 

Here is the size on disk change (if you want to see other assemblies, let me know)
```
System.Private.Corelib size
  Before:     9,296 KB on disk
  After :     9,262 KB on disk

Savings: 34 KB
```

All perf tests were run warm by running for 2 iterations and discarding the results. The baseline here is master before my changes.

The high level summary is I ran startup tests for WPF, ASP.Net and Powershell and all results were within the noise - no gains, but no losses either.

Here is the timing for the WPF app Bill provided:
```
    Before:   
              601 msec
              582 msec
              620 msec
              587 msec
              575 msec
              Average: 593 msec
    After :   
              611 msec
              594 msec
              582 msec
              588 msec
              583 msec
              Average: 592 msec

```

I tested powershell using instructions from @kouvel, grabbing the latest powershell build and replacing the coreclr binaries with those built from master. The script is as follows (it measures the average time to start up powershell and exit):
```
$iterationCount = 32
$totalDuration = 0.0

1..$iterationCount | % {
  $duration = (Measure-Command { .\pwsh\pwsh.exe -noprofile -c exit }).TotalMilliseconds
  $duration.ToString("0.000")
  $totalDuration += $duration
}

"Average: " + ($totalDuration / $iterationCount).ToString("0.000")
```
Here is the timings:
```
Baseline powershell:
  309.215
  293.471
  295.329
  298.485
  290.840
  307.471
  302.760
  309.857
  302.411
  314.054
  302.502
  295.069
  332.554
  305.581
  291.474
  304.776
  299.794
  317.381
  293.785
  295.340
  302.556
  292.813
  305.382
  293.320
  292.424
  301.300
  293.092
  315.784
  292.458
  297.204
  292.774
  293.598
  Average: 301.089

With my changes:
  300.420
  294.193
  297.597
  290.947
  301.054
  313.396
  292.750
  299.602
  292.826
  305.875
  295.943
  293.836
  293.632
  292.098
  301.508
  293.552
  297.416
  296.621
  314.508
  295.243
  297.459
  325.361
  307.380
  303.980
  302.602
  290.661
  325.002
  294.576
  295.850
  292.906
  316.386
  294.502
  Average: 300.303
```
To test ASP.NET startup I used MusicStore, although none of the automated ways to run it appear to be working so I had to manually build it.

```

MusicStore baseline:
  D:\jumpstamp_binaries\musicstore_bin>CoreRun.exe MusicStore.dll
  Server started in 704ms

  Starting request to http://localhost:5000
  Response: OK
  Request took 442ms

  Cold start time (server start + first request time): 1146ms


  Batch 0: running 1001 requests
  Steadystate min response time: 2.00ms
  Steadystate max response time: 237.00ms
  Steadystate average response time: 3.45ms
  Steadystate median response time: 3.04ms
  Batch 1: running 1001 requests
  Steadystate min response time: 2.00ms
  Steadystate max response time: 5.00ms
  Steadystate average response time: 2.89ms
  Steadystate median response time: 2.78ms
  Batch 2: running 1001 requests
  Steadystate min response time: 2.00ms
  Steadystate max response time: 4.00ms
  Steadystate average response time: 2.87ms
  Steadystate median response time: 2.78ms
  Batch 3: running 1001 requests
  Steadystate min response time: 2.00ms
  Steadystate max response time: 144.00ms
  Steadystate average response time: 3.06ms
  Steadystate median response time: 2.83ms
  Batch 4: running 1001 requests
  Steadystate min response time: 2.00ms
  Steadystate max response time: 4.00ms
  Steadystate average response time: 2.86ms
  Steadystate median response time: 2.77ms

MusicStore with Changes:
  D:\jumpstamp_binaries\musicstore_bin>CoreRun.exe MusicStore.dll
  Server started in 707ms

  Starting request to http://localhost:5000
  Response: OK
  Request took 449ms

  Cold start time (server start + first request time): 1156ms


  Batch 0: running 1001 requests
  Steadystate min response time: 2.00ms
  Steadystate max response time: 37.00ms
  Steadystate average response time: 3.34ms
  Steadystate median response time: 3.09ms
  Batch 1: running 1001 requests
  Steadystate min response time: 2.00ms
  Steadystate max response time: 221.00ms
  Steadystate average response time: 3.11ms
  Steadystate median response time: 2.81ms
  Batch 2: running 1001 requests
  Steadystate min response time: 2.00ms
  Steadystate max response time: 4.00ms
  Steadystate average response time: 2.85ms
  Steadystate median response time: 2.74ms
  Batch 3: running 1001 requests
  Steadystate min response time: 2.00ms
  Steadystate max response time: 5.00ms
  Steadystate average response time: 2.83ms
  Steadystate median response time: 2.73ms
  Batch 4: running 1001 requests
  Steadystate min response time: 2.00ms
  Steadystate max response time: 163.00ms
  Steadystate average response time: 3.01ms
  Steadystate median response time: 2.77ms
```